Implement rating count using cloud functions
