+++ functions/package.json
@@ -11,10 +11,12 @@
   "engines": {
     "node": "16"
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "dependencies": {
+    "eslint": "^8.33.0",
     "firebase-admin": "^10.0.2",
-    "firebase-functions": "^3.18.0"
+    "firebase-functions": "^4.2.1",
+    "prettier": "^2.8.4"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.2.0"
