+++ functions/src/index.js
@@ -1,0 +1,1 @@
+exports.ratingCount = require("./rating_count/rating_count");
