+++ lib/src/constants/mocks/courses.dart
@@ -13,8 +13,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/240x135/706020_6d7c_3.jpg',
     instructorName: 'Mark Wahlbeck',
-    rating: 4.5,
-    ratingCount: 785,
     price: 59.99,
     salePrice: 11.99,
     youWillLearn: _youWillLearn,
@@ -27,8 +25,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/240x135/900464_a431_22.jpg',
     instructorName: 'Morteza Kordi',
-    rating: 4.4,
-    ratingCount: 4389,
     price: 29.99,
     youWillLearn: _youWillLearn,
     category: 'mobile development',
@@ -40,8 +36,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/240x135/919872_ed54_6.jpg',
     instructorName: 'Tim Buchalka',
-    rating: 4.3,
-    ratingCount: 10344,
     price: 59.99,
     salePrice: 19.99,
     youWillLearn: _youWillLearn,
@@ -54,8 +48,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/240x135/921168_7646_2.jpg',
     instructorName: 'Anna Purrna Argawal Sidhartha',
-    rating: 4.3,
-    ratingCount: 10344,
     price: 59.99,
     salePrice: 19.99,
     youWillLearn: _youWillLearn,
@@ -68,8 +60,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/240x135/919872_ed54_6.jpg',
     instructorName: 'Tim Buchalka',
-    rating: 4.3,
-    ratingCount: 10344,
     price: 59.99,
     salePrice: 19.99,
     youWillLearn: _youWillLearn,
@@ -82,8 +72,6 @@ final courses = <LCourse>[
         'From Beginner to iOS App Developer with Just One Course! Fully Updated with a Comprehensive Module Dedicated to SwiftUI!',
     thumbnail: 'https://img-b.udemycdn.com/course/480x270/1829188_efd3_5.jpg',
     instructorName: 'Anushka Madusanka',
-    rating: 4.2,
-    ratingCount: 24279,
     price: 89.0,
     salePrice: 9.9,
     youWillLearn: _youWillLearn,
