+++ lib/src/domain_manager.dart
@@ -9,6 +9,7 @@ import 'features/cart/data/local/local__cart_repository.dart';
 import 'features/cart/data/remote/remote_cart_repository.dart';
 import 'features/course/data/course_repository.dart';
 import 'features/profile/data/user_repository.dart';
+import 'features/rating_count/data/rating_count_repository.dart';
 import 'features/reviews/data/review_repository.dart';
 import 'features/wishlist/data/wishlist_repository.dart';
 
@@ -50,5 +51,9 @@ class DomainManager {
 
   final reviewRepositoryProvider = Provider<ReviewRepository>((ref) {
     return FirestoreReviewRepository(FirebaseFirestore.instance);
+  });
+
+  final ratingCountRepositoryProvider = Provider<RatingCountRepository>((ref) {
+    return FirestoreRatingCountRepository(FirebaseFirestore.instance);
   });
 }
